LocationImages changes, Beacon Card xml
