+++ Android/AdventourAndroidVersion/app/src/main/java/com/adventour/android/LocationImages.java
@@ -1,0 +1,17 @@
+package com.adventour.android;
+
+import android.graphics.Bitmap;
+
+public class LocationImages
+{
+    Bitmap locationOne;
+    Bitmap locationTwo;
+    Bitmap locationThree;
+
+    public void locationImages(Bitmap locationOne, Bitmap locationTwo, Bitmap locationThree)
+    {
+        this.locationOne = locationOne;
+        this.locationTwo = locationTwo;
+        this.locationThree = locationThree;
+    }
+}
