+++ Android/AdventourAndroidVersion/app/src/main/res/layout/beacon_card_layout.xml
@@ -36,6 +36,7 @@
             android:clickable="true"
             android:layout_width="30dp"
             android:layout_height="30dp"
+            android:layout_marginRight="5dp"
             android:src="@drawable/ic_heart_line_icon"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
@@ -63,15 +64,16 @@
 
         <TextView
             android:id="@+id/likesTextView"
-            android:layout_width="80dp"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:ellipsize="end"
             android:maxLines="1"
-            android:textColor="@color/black"
+            android:textColor="@color/red_variant"
             android:textSize="18dp"
             android:textStyle="bold"
             app:layout_constraintBottom_toTopOf="@+id/beaconTitleTextView"
-            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="@id/likeImageView"
+            app:layout_constraintEnd_toEndOf="@id/likeImageView"
             app:layout_constraintTop_toBottomOf="@+id/likeImageView" />
 
         <TextView
